Simple Sample Player
Simple Sample player for Max4Live
With phase stacking and cropping issues in all modes of the default samplers Simpler and Sampler, and Tonal changes in Impulse I have decided to create a Sampler that you press any key and it just plays the sample as is.

Load on a MIDI Track Drag the sample on and send press any key and no changes to the sound will be made other than just playing it. Add your own volume control if needed,

They dont stack by the way yet, so this is expected behaviour for the moment, but expected changes will be that I will release one with and octave switch and get stacking working.

Expected use is for when you render a sound ifor example a Kick and then want to trigger it on the beat the exact same every time.